backend/requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
gpiod==2.3.0
gpiodevice==0.0.5
idna==3.10
importlib_metadata==8.6.1
inky==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.0.2
pillow==11.1.0
requests==2.32.3
smbus2==0.5.0
spidev==3.6
urllib3==2.3.0
Werkzeug==3.1.3
zipp==3.21.0
